PaxWatchlistLink Services, Repos, and Dependencies
-New Dependency Apache commons-text
-Added constructor to paxwatchlistlinkvo and modified some field datatypes
-Modified PaxWatchlistLink model matchpercentage to float, requires DB hibernate4:export
-Added MatchingService which uses Jarawinkler and Double Metaphone for string matching.

+++ gtas-parent/gtas-commons/src/main/java/gov/gtas/model/PaxWatchlistLink.java
@@ -32,7 +32,7 @@ public class PaxWatchlistLink {
     private Long id;  
     
     @Column(name= "percent_match")
-    private int percentMatch;
+    private float percentMatch;
     
     @Column(name= "last_run_timestamp")
     @Temporal(TemporalType.TIMESTAMP)
@@ -54,11 +54,11 @@ public class PaxWatchlistLink {
 		this.id = id;
 	}
 
-	public int getPercentMatch() {
+	public float getPercentMatch() {
 		return percentMatch;
 	}
 
-	public void setPercentMatch(int percentMatch) {
+	public void setPercentMatch(float percentMatch) {
 		this.percentMatch = percentMatch;
 	}
